Updated all working space and its dependencies Refactored and optimized some code Thanks to GitHub Copilot and UMA's license 💙💙💙
