+++ app/src/main/java/star/key13bb/emerald/ContextHandler.kt
@@ -1,0 +1,23 @@
+package star.key13bb.emerald
+
+import android.app.Application
+import android.content.Context
+
+class MyApplication : Application() {
+    init {
+        instance = this
+    }
+
+    companion object {
+        private var instance: MyApplication? = null
+
+        fun getContext(): Context {
+            return instance!!.applicationContext
+        }
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        // Initialize other components if needed
+    }
+}
