+++ app/src/main/java/star/key13bb/emerald/ui/portrait/ExpandMapsScreen.kt
@@ -1,0 +1,14 @@
+package star.key13bb.emerald.ui.portrait
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import star.key13bb.emerald.ui.screens.DefaultScreen
+
+@Composable
+fun ExpandMapsScreen () {
+    Box(modifier = Modifier.fillMaxSize()) {
+        DefaultScreen()
+    }
+}
