--- app/src/main/java/star/kiko13bb/emerald/ui/portrait/ExpandMapsScreen.kt
@@ -1,14 +1,0 @@
-package star.kiko13bb.emerald.ui.portrait
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import star.kiko13bb.emerald.ui.screens.DefaultScreen
-
-@Composable
-fun ExpandMapsScreen () {
-    Box(modifier = Modifier.fillMaxSize()) {
-        DefaultScreen()
-    }
-}
