+++ app/src/main/proto/scheme.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-option java_package = "star.kiko13bb.emerald";
+option java_package = "star.key13bb.emerald";
 option java_multiple_files = true;
 
 // IDs must be > 0
