+++ gradle.properties
@@ -10,11 +10,11 @@
 # This option should only be used with decoupled projects. For more details, visit
 # https://developer.android.com/r/tools/gradle-multi-project-decoupled-projects
 # org.gradle.parallel=true
-#Mon Aug 12 16:16:34 CEST 2024
+#Wed Oct 16 23:21:38 CEST 2024
 android.nonTransitiveRClass=true
 android.suppressUnsupportedCompileSdk=35
 android.useAndroidX=true
 kotlin.code.style=official
 org.gradle.caching=true
 org.gradle.configuration-cache=true
-org.gradle.jvmargs=-Xmx1024M -Dkotlin.daemon.jvm.options\="-Xmx1024M" -Dfile.encoding\=UTF-8
+org.gradle.jvmargs=-Xmx1536M -Dkotlin.daemon.jvm.options\="-Xmx1536M" -Dfile.encoding\=UTF-8
