+++ gradle/libs.versions.toml
@@ -1,18 +1,18 @@
 [versions]
 datastore-preferences-core-version = "1.1.1"
-androidx-ui = "1.8.0-alpha01"
+androidx-ui = "1.8.0-alpha04"
 datastore-preferences = "1.1.1"
 gtfs-validator-model = "5.0.0"
 kotlinx-coroutines-android = "1.9.0"
-lifecycle = "2.9.0-alpha02"
+lifecycle = "2.9.0-alpha05"
 
 [libraries]
 
-androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version = "2024.09.01" }
-firebase-bom = { module = "com.google.firebase:firebase-bom", version = "33.3.0" }
+androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version = "2024.10.00" }
+firebase-bom = { module = "com.google.firebase:firebase-bom", version = "33.4.0" }
 
 androidx-activity-compose = { group = "androidx.activity", name = "activity-compose", version = "1.10.0-alpha02" }
-androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version = "1.15.0-alpha02" }
+androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version = "1.15.0-rc01" }
 androidx-datastore-core = { module = "androidx.datastore:datastore-core", version.ref = "datastore-preferences-core-version" }
 androidx-datastore-preferences = { group = "androidx.datastore", name = "datastore-preferences", version.ref = "datastore-preferences" }
 
@@ -23,8 +23,8 @@ androidx-lifecycle-viewmodel-compose = { module = "androidx.lifecycle:lifecycle-
 androidx-lifecycle-viewmodel-ktx = { module = "androidx.lifecycle:lifecycle-viewmodel-ktx", version.ref = "lifecycle" }
 androidx-lifecycle-viewmodel-savedstate = { module = "androidx.lifecycle:lifecycle-viewmodel-savedstate", version.ref = "lifecycle" }
 
-androidx-material3 = { group = "androidx.compose.material3", name = "material3", version = "1.3.0" }
-androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version = "2.8.0" }
+androidx-material3 = { group = "androidx.compose.material3", name = "material3", version = "1.4.0-alpha02" }
+androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version = "2.9.0-alpha01" }
 androidx-preference-ktx = { group = "androidx.preference", name = "preference-ktx", version = "1.2.1" }
 androidx-ui = { group = "androidx.compose.ui", name = "ui", version.ref = "androidx-ui" }
 androidx-ui-graphics = { group = "androidx.compose.ui", name = "ui-graphics", version.ref = "androidx-ui" }
@@ -32,21 +32,21 @@ androidx-ui-tooling-preview = { group = "androidx.compose.ui", name = "ui-toolin
 
 datastore-preferences-core = { module = "androidx.datastore:datastore-preferences-core", version.ref = "datastore-preferences-core-version" }
 
-firebase-analytics = { module = "com.google.firebase:firebase-analytics", version = "22.1.0" }
-firebase-crashlytics = { module = "com.google.firebase:firebase-crashlytics", version = "19.1.0" }
+firebase-analytics = { module = "com.google.firebase:firebase-analytics", version = "22.1.2" }
+firebase-crashlytics = { module = "com.google.firebase:firebase-crashlytics", version = "19.2.0" }
 
 gtfs-realtime-bindings = { module = "org.mobilitydata:gtfs-realtime-bindings", version = "0.0.8" }
 gtfs-validator-model = { module = "org.mobilitydata.gtfs-validator:gtfs-validator-model", version.ref = "gtfs-validator-model" }
 
 kotlinx-coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "kotlinx-coroutines-android" }
-onebusaway-gtfs = { module = "org.onebusaway:onebusaway-gtfs", version = "3.2.3" }
-protobuf-kotlin = { module = "com.google.protobuf:protobuf-java", version = "3.25.4" }
+onebusaway-gtfs = { module = "org.onebusaway:onebusaway-gtfs", version = "3.2.4" }
+protobuf-kotlin = { module = "com.google.protobuf:protobuf-java", version = "4.29.0-RC1" }
 
 androidx-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-manifest", version.ref = "androidx-ui" }
 androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling", version.ref = "androidx-ui" }
 
 
 [plugins]
-android-application = { id = "com.android.application", version = "8.6.0" }
+android-application = { id = "com.android.application", version = "8.8.0-alpha06" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version = "1.9.0" }
 dependency-sorter = { id = "com.squareup.sort-dependencies", version = "0.8"}
