+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,7 @@
-#Tue Sep 10 16:04:54 CEST 2024
+#Wed Oct 16 23:33:14 CEST 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.10.1-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.11-milestone-1-bin.zip
 networkTimeout=10000
 validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
